requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-common-compat==1.6.0
apache-airflow-providers-apache-spark==5.1.1
pandas==2.1.4
loguru==0.7.3
requests==2.32.0
boto3==1.34.34
openpyxl==3.1.2
pyarrow==14.0.1
delta-spark==3.3.0